requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
praw==7.1.0
prawcore==1.5.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
PySocks==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
twilio==6.45.3
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
wrapt==1.12.1
